refactor: reorganized file structure of logging lib
