+++ src/src/cpp/TPG/SConscript
@@ -3,6 +3,6 @@ env.Library('TPG', ['TPG.cc', 'instruction.cc', 'RegisterMachine.h',
                     'RegisterMachine.cc','point.cc','team.cc','misc.cc'],
                     LIBS = ['boost_iostreams','eigen3', 'logging'],
                     LIBPATH = ['/lib/x86_64-linux-gnu/', '../logging'],
-                    CPPPATH=['.','../environments','../experiments','/usr/include/eigen3', '../logging'])
+                    CPPPATH=['.','../environments','../experiments','/usr/include/eigen3', '../logging', '../logging/core', '../logging/loggers', '../logging/metrics', '../logging/storage'])
 
 env.Append(CPPPATH = ['../experiments', '../logging'])
