+++ src/src/cpp/logging/core/event_dispatcher.h
