+++ src/src/cpp/logging/core/event_dispatcher.tpp
@@ -1,4 +1,4 @@
-#include <event_dispatcher.h>
+#include "event_dispatcher.h"
 
 template <typename MetricsType>
 EventDispatcher<MetricsType>& EventDispatcher<MetricsType>::instance() {
