+++ src/src/cpp/logging/core/event_types.h
