+++ src/src/cpp/logging/loggers/mta/mta_logger.h
@@ -1,6 +1,6 @@
-#include <csv_storage.h>
-#include <event_dispatcher.h>
-#include <mta_metrics.h>
+#include "storage/csv_storage.h"
+#include "core/event_dispatcher.h"
+#include "metrics/mta/mta_metrics.h"
 
 
 class MTALogger {
