+++ src/src/cpp/logging/metrics/mta/mta_metrics.cc
@@ -1,5 +1,5 @@
-#include <mta_metrics.h>
-#include <mta_metrics_builder.h>
+#include "mta_metrics.h"
+#include "mta_metrics_builder.h"
 
 MTAMetrics::MTAMetrics(const MTAMetricsBuilder& builder)
     : team_id(builder.get_team_id()),
