+++ src/src/cpp/logging/metrics/mta/mta_metrics_builder.cc
@@ -1,5 +1,5 @@
-#include <mta_metrics_builder.h>
-#include <mta_metrics.h>
+#include "mta_metrics_builder.h"
+#include "mta_metrics.h"
 
 MTAMetricsBuilder& MTAMetricsBuilder::with_generation(long generation) {
     this->generation = generation;
