+++ src/src/cpp/logging/metrics/tms/timing_metrics_builder.cc
@@ -1,5 +1,5 @@
-#include <timing_metrics_builder.h>
-#include <timing_metrics.h>
+#include "timing_metrics_builder.h"
+#include "timing_metrics.h"
 
 TimingMetricsBuilder& TimingMetricsBuilder::with_generation(int generation) {
     this->generation = generation;
