+++ src/src/cpp/logging/storage/mta/mta_storage.cc
